# Optimize cache.Set operation